+++ src/test/java/com/translantik/runners/CukesRunner.java
@@ -13,7 +13,7 @@ import org.junit.runner.RunWith;
         features = "src/test/resources",
         glue = "com/translantik/step_definitions",
         dryRun = false,
-        tags = "@TRAN-252"
+        tags = "smoke"
 )
 public class CukesRunner {
