+++ src/test/java/com/translantik/step_definitions/VehicleCostStepDef.java
@@ -142,4 +142,18 @@ public class VehicleCostStepDef {
         Assert.assertFalse(new VehicleCostPage().vcGridSettingCheckBoxes.get(0).isSelected());
         Assert.assertFalse(new VehicleCostPage().vcGridSettingCheckBoxes.get(1).isSelected());
     }
+
+    //US_029 By Ferruh
+
+    @When("the user clicks Create Vehicle Costs button")
+    public void the_user_clicks_Create_Vehicle_Costs_button() {
+        new VehicleCostPage().createVehicleCostsButton.click();
+    }
+
+    @Then("the user should be able to click Save And Close button")
+    public void the_user_should_be_able_to_click_Save_And_Close_button() {
+        new VehicleCostPage().saveAndCloseButton.click();
+        //Driver.get().findElement(By.xpath("//button[@type='submit']")).click();
+
+    }
 }
